Replace corrupted symlink with standard code file
Previously, `src/GenericRTC.h` was a symbolic link for which the target path was the entire text of the code that was
intended to be the content of the file.

This invalid target path caused compilation of the library to fail:

```
DS3232RTC.h:14:10: fatal error: GenericRTC.h: No such file or directory
 #include <GenericRTC.h>
          ^~~~~~~~~~~~~~
compilation terminated.
```

It also caused the Library Manager indexer to be unable to check out the release:

```
2025/06/25 12:09:43 Checking out tag: 3.1.0
2025/06/25 12:09:43 Error syncing library: error checking out repo: failed to get repository to clean state
```

Even when functional, use of symlinks in Arduino libraries is generally problematic:

- Releases containing a symlink are rejected by the Library Manager indexer.
- Even with a manual installation, they are unfriendly to Windows users since symlinks are disabled in Windows by
  default.

So the chosen solution is to replace the symlink with a normal code file.